Internal: minor changes
- Adding an error message if neither POT_RELEASE nor POT_DEBUG is defined
- Removing a useless C-style cast

+++ trunk/src/PotatoEngine/DNACollectorChecker.cpp
@@ -277,7 +277,7 @@ namespace Pot
 
 				std::ostringstream oss;
 				oss << "DNACollector::checkIntegrity: ";
-				oss << "the organism " << (long)containerIt->first << " is " << (alive ? "alive" : "dead");
+				oss << "the organism " << containerIt->first << " is " << (alive ? "alive" : "dead");
 				oss << " but some DNAs referencing it are " << (alive ? "dead" : "alive");
 				handleError(checkDNAsStatus(t, alive), oss.str().c_str()); /* O(DnaC) */
 			}
